Bump Gradle to 9.2.1
It turned out that the shadow plugin version 9.3.0, which has Java 17 as the baseline, pulls in Kotlin 2.2 as well, which is incompatible with Gradle 8.x (demanding exactly Kotlin 2.0). The Kotlin requirements for Gradle DSL scripts are hard-coded in the Gradle code base.

This change bumps Gradle to 9.2.1 to accomodate incompatible Gradle plugin versions.

We could downgrade the shadow plugin, but OTOH it's okay to bump Gradle in NesQuEIT to 9.2.1, with some slight modifications to the Iceberg build customization https://github.com/snazy/iceberg/tree/iceberg-nesqueit, which is pulled in by all NesQuEIT CI invocations.

Supersedes #813