Remove no longer necessary complicated stuff
Initially we had to deal with conflicting Maven group-IDs for Nessie, at the time when all Nessie modules were published using the same group ID, which was a problem in NesQuEIT for Nessie modules that depend on Iceberg.
Since then, we have migrated all modules to `org.projectnessie.nessie-integrations` to decouple that.

NesQuEIT had to live with the ugly workarounds for the included builds for a long time.
It's time to remove these no longer necessary stuff.